Extracted appointment implementations into their own compilation unit
