+++ src/appointment_ranges/appointment_ranges_lib.cpp
@@ -11,24 +11,6 @@ bool operator==(const Range& lhs, const Range& rhs)
         lhs.pattern == rhs.pattern;
 }
 
-bool operator<(const Appointment& lhs, const Appointment& rhs)
-{
-    if (lhs.startMinute != rhs.startMinute)
-    {
-        return lhs.startMinute < rhs.startMinute;
-    }
-    else if(lhs.endMinute != lhs.startMinute)
-    {
-        // Looks like a bug! Bug we actually want the larger ranges
-        // to sort before the shorter ranges, all other things being equal.
-        return lhs.endMinute > rhs.endMinute;
-    }
-    else /* (lhs.state != rhs.state)*/
-    {
-        return lhs.state < rhs.state;
-    }
-}
-
 Pattern getPatternFromState(FreeBusy state)
 {
     switch(state)
@@ -40,16 +22,6 @@ Pattern getPatternFromState(FreeBusy state)
     default:
         return Pattern::Empty;
     }
-}
-
-bool Appointment::startsAfter(int minute) const
-{
-    return startMinute > minute;
-}
-
-bool Appointment::endsAfter(int minute) const
-{
-    return endMinute > minute;
 }
 
 void Ranges::addRange(Range&& newRange)
